pyproject.toml
[tool.poetry]
name = "nomad-django-project"
version = "0.1.0"
description = ""
authors = ["이원빈 <2wonbin@kakao.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.1"
pillow = "^10.4.0"
djangorestframework = "3.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
